Исправила недочеты и баги
